pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devil-api"
version = "0.1.0"
description = "fastAPI for devil servers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "Mateusz Hajder"},
]

dependencies = [
    "fastapi>=0.116",
    "uvicorn[standard]>=0.35",
    "pydantic>=2.11,<3",
    "python-dotenv>=1.1.1,<2",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pre-commit",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["app/main"]

[project.urls]
Homepage = "https://github.com/devil-imps/devil-api"
Repository = "https://github.com/devil-imps/devil-api"
Issues = "https://github.com/devil-imps/devil-api/issues"

[tool.uv]
default-groups = []

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11+.
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "S",  # flake8-bandit
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "LOG", # flake8-logging
    "RUF", # Ruff-specific rules
]

ignore = [
    "E501",  # Line too long, handled by formatter
    "S101",  # Use of assert (common in tests)
    "S104",  # Possible binding to all interfaces
    "S105",  # Possible hardcoded password
    "S106",  # Possible hardcoded password
    "S107",  # Possible hardcoded password
    "S108",  # Probable insecure usage of temp file/directory
    "T201",  # print found (we use print statements for CLI output)
    "B008",  # Do not perform function calls in argument defaults
    "N806",  # Variable name should be lowercase
    "SIM118", # Use dict.get() instead of conditional assignment
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106", "S107"]  # Allow assert and hardcoded passwords in tests
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files

[tool.ruff.lint.isort]
known-first-party = ["devil_api"]
force-single-line = true
